DASproekt/requirements.txt
confluent_kafka
yfinance
pandas
pytz
psycopg2-binary
SQLAlchemy
Flask
prophet
pyspark
